+++ CHANGES.md
@@ -10,6 +10,7 @@ New:
 - Added LV2 support (#906).
 - Added `string.nth` (#970).
 - Added `string.binary.to_int` (#970).
+- Added `string.hex_of_int`.
 - Added `file.ls` (#1011).
 - Added native id3v2 tag parser, as well as associated function
   `file.mp3.metadata`, `file.mp3.parse_apic` and `file.cover` (#987).